% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruits1}
\alias{fruits1}
\title{Aminoacid concentration in fruit juice samples}
\format{
A data frame with 36 rows and 11 variables:

\describe{

\item{LYS}{lysine}
\item{ARG}{arginine}
\item{ASP}{aspartic acid}
\item{SER}{serine}
\item{GLU}{glutamine acid}
\item{PRO}{proline}
\item{GLY}{glycine}
\item{ALA}{alanine}
\item{VAL}{valine}
\item{PHA}{phenyl alanine}
\item{GABA}{gamma-amino butric acid}
}
}
\usage{
fruits1
}
\description{
Simulated data set of Phase I samples of pure fruit juice.
Original concentrations are assumed to be measured in \eqn{\mu}g per standard volume.
Variables are standardized.
}
\references{
Fuchs, S. & Kenett, R. S. (1998).
"Multivariate quality control: theory and applications".
Chapman and Hall/CRC.
}
\seealso{
\code{\link[=fruits2]{fruits2()}}, \code{\link[=fruits2na]{fruits2na()}}
}
\keyword{datasets}
