% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruits2na}
\alias{fruits2na}
\title{Concentración de 11 aminoácidos en jugo de fruta puro}
\format{
A data frame with 30 rows and 11 variables:

\strong{uuu}: hola

\describe{

\item{LYS}{lisina}
\item{ARG}{arginina}
\item{ASP}{ácido aspártico}
\item{SER}{serina}
\item{GLU}{ácido glutámico}
\item{PRO}{prolina}
\item{GLY}{glicina}
\item{ALA}{alanina}
\item{VAL}{valina}
\item{PHA}{fenilalanina}
\item{GABA}{ácido gamma-aminobutírico}
}
}
\usage{
fruits2na
}
\description{
Conjunto de datos simulado. Las concentraciones se asumen
medidas en microgramos por volumen estándar.
Los valores faltantes fueron generados MCAR.
}
\keyword{datasets}
