% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frutasI}
\alias{frutasI}
\title{Concentración de 11 aminoácidos en jugo de fruta puro}
\format{
A data frame with 36 rows and 11 variables:

\strong{uuu}: hola

\describe{

\item{LYS}{lisina}
\item{ARG}{arginina}
\item{ASP}{ácido aspártico}
\item{SER}{serina}
\item{GLU}{ácido glutámico}
\item{PRO}{prolina}
\item{GLY}{glicina}
\item{ALA}{alanina}
\item{VAL}{valina}
\item{PHA}{fenilalanina}
\item{GABA}{ácido gamma-aminobutírico}
}
}
\usage{
frutasI
}
\description{
Conjunto de datos simulado. Las concentraciones se asumen
medidas en microgramos por volumen estándar.
}
\keyword{datasets}
