% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficoSPE.R
\name{graficoSPE}
\alias{graficoSPE}
\title{Gr치fico SPE}
\usage{
graficoSPE(x, alfa, autov, A)
}
\arguments{
\item{x}{salida de score_imp()}

\item{alfa}{nivel de significacion}

\item{autov}{autovalores}

\item{A}{cantidad de componentes elegidas}
}
\value{
gr치fico SPE
}
\description{
Esta funci칩n genera el gr치fico SPE
}
\details{
agregar detalles
}
