% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficoT2.R
\name{graficoT2}
\alias{graficoT2}
\title{Gráfico T^2}
\usage{
graficoT2(x, alfa)
}
\arguments{
\item{x}{salida de la función cmr_tsr()}

\item{alfa}{nivel de significacion}
}
\value{
devuelve una lista con los siguientes elementos: i) gráfico T^2 generado, ii) proporción de observaciones fuera de control, iii) ID de observaciones fuera de control
}
\description{
Esta función genera el gráfico T^2
}
\details{
agregar detalles
}
