% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_imp.R
\name{score_imp}
\alias{score_imp}
\title{Imputación de Scores CMR y TSR}
\usage{
score_imp(datos, pesos, autov, A, metodo)
}
\arguments{
\item{datos}{vector o matriz de datos numericos que pueden tener missing values}

\item{pesos}{matriz de pesos que sale del PCA}

\item{autov}{autovalores}

\item{A}{cantidad de componentes elegidas}

\item{metodo}{CMR o TSR}
}
\value{
devuelve una lista con scores y errores calculados
}
\description{
Esta función calcula scores usando el metodo CMR (KDR) o TSR
}
\details{
agregar detalles
}
\examples{

}
