% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe.chart.R
\name{spe.chart}
\alias{spe.chart}
\title{SPE Control Chart}
\usage{
spe.chart(scores, alpha = 0.005)
}
\arguments{
\item{scores}{output from \code{score_imp()} function.}

\item{alpha}{significance level used for computing the Upper Control Limit (UCL).
Defaults to 0.005.}
}
\value{
\code{spe.chart} returns a list containing the following components:
\describe{
\item{Plot}{ggplot object with SPE control chart}
\item{poc}{Proportion of out-of-control samples}
\item{out}{ID of out-of-control samples}
\item{UCL}{value of the Upper Control Limit}
}
}
\description{
This function plots the SPE Control Chart.
It also returns the proportion of out-of-control samples and their position,
along with the value of the computed Upper Control Limit.
}
\details{
The calculation of the Upper Control Limit is based on Jackson & Mudholkar (1979).
}
\examples{
#Score imputation with CMR method
scores <- score_imp(
  data1 = fruits1,
  data2 = fruits2na,
  A = 4,
  method = "CMR"
  )
spe.chart(scores = scores, alpha = 0.005)
}
\seealso{
\code{\link[=score_imp]{score_imp()}}
}
