% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2.chart.R
\name{t2.chart}
\alias{t2.chart}
\title{\eqn{T^2} Control Chart of PCs}
\usage{
t2.chart(scores, alpha = 0.005, phase = 2, m)
}
\arguments{
\item{scores}{output from \code{score_imp()} function.}

\item{alpha}{significance level used for computing the Upper Control Limit (UCL).
Defaults to 0.005.}

\item{phase}{1 or 2. Defaults to 2.}

\item{m}{number of Phase I samples.}
}
\value{
\code{t2.chart} returns a list containing the following components:
\describe{
\item{Plot}{ggplot object with \eqn{T^2} Control Chart of PCs}
\item{poc}{Proportion of out-of-control samples}
\item{out}{ID of out-of-control samples}
\item{UCL}{Value of the Upper Control Limit}
}
}
\description{
This function plots the \eqn{T^2} Control Chart of PCs.
It also returns the proportion of out-of-control samples and their position,
along with the value of the computed Upper Control Limit.
}
\details{
The computation of the Upper Control Limit (UCL)
depends on the \code{phase} argument. The chart is implemented for samples
of size 1.
}
\examples{
#Score imputation with CMR method
scores <- score_imp(
  data1 = fruits1,
  data2 = fruits2na,
  A = 4,
  method = "CMR"
  )
#Phase I data has 36 observatios
t2.chart(scores = scores, alpha = 0.005, phase = 2, m = 36)
}
\references{
MacGregor, J. F., & Kourti, T. (1995).
"Statistical process control of multivariate processes."
Control engineering practice, 3(3), 403-414.

Wold, S., Geladi, P., Esbensen, K., & Ohman, J. (1987).
"Multi‐way principal components‐and PLS‐analysis"
Journal of chemometrics, 1(1), 41-56.
}
\seealso{
\code{\link[=score_imp]{score_imp()}}
}
